LPS-70711 Correct names of REST API modules (JAX-RS -> REST)
