--- adaptive-media-image-jax-rs-test/bnd.bnd
@@ -1,10 +1,0 @@
-Bundle-Name: Liferay Adaptive Media Image JAX-RS Configuration Test
-Bundle-SymbolicName: com.liferay.adaptive.media.image.jaxrs.configuration.test
-Bundle-Version: 1.0.0
-Import-Package:\
-	!com.google.gson.*,\
-	\
-	!org.glassfish.jersey.client.*,\
-	*
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Adaptive Media
