--- adaptive-media-image-jax-rs-test/src/testIntegration/resources/bnd.bnd
@@ -1,7 +1,0 @@
-Bundle-Activator: com.liferay.adaptive.media.image.jaxrs.media.test.internal.activator.configuration.TestBundleActivator
-Bundle-Name: Liferay Adaptive Media Image JAX-RS Media Test
-Bundle-SymbolicName: com.liferay.adaptive.media.image.jaxrs.media.test
-Bundle-Version: 1.0.0
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Adaptive Media
-Private-Package: com.liferay.adaptive.media.image.jaxrs.media.test.*
