--- adaptive-media-image-jax-rs/bnd.bnd
@@ -1,16 +1,0 @@
-Bundle-Name: Liferay Adaptive Media Image JAX-RS
-Bundle-SymbolicName: com.liferay.adaptive.media.image.jaxrs
-Bundle-Version: 1.0.0
-ConfigurationPath: /configuration
-Import-Package:\
-	!com.fasterxml.jackson.*,\
-	\
-	!com.google.gson,\
-	\
-	!javassist.*,\
-	\
-	!org.apache.commons.vfs2.*,\
-	*
-Include-Resource: configuration=src/main/resources/configuration
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Adaptive Media
