--- adaptive-media-image-jax-rs/src/main/resources/configuration/com.liferay.portal.remote.rest.extender.configuration.RestExtenderConfiguration-adaptivemediajaxrs
@@ -1,2 +1,0 @@
-contextPaths=/adaptive-media-rest
-jaxRsApplicationFilterStrings=(component.name=com.liferay.adaptive.media.image.jaxrs.internal.ImageAdaptiveMediaJaxRsApplication)
