+++ adaptive-media-image-rest-test/bnd.bnd
@@ -1,0 +1,10 @@
+Bundle-Name: Liferay Adaptive Media Image REST Test
+Bundle-SymbolicName: com.liferay.adaptive.media.image.rest.test
+Bundle-Version: 1.0.0
+Import-Package:\
+	!com.google.gson.*,\
+	\
+	!org.glassfish.jersey.client.*,\
+	*
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Adaptive Media
