+++ adaptive-media-image-rest-test/build.gradle
