+++ adaptive-media-image-rest-test/libs/com.liferay.document.library.demo.data.creator.api-1.1.1.jar
