+++ adaptive-media-image-rest-test/libs/com.liferay.document.library.demo.data.creator.impl-1.0.5.jar
