+++ adaptive-media-image-rest-test/libs/com.liferay.users.admin.demo.data.creator.api-1.0.3.jar
