+++ adaptive-media-image-rest-test/libs/com.liferay.users.admin.demo.data.creator.impl-1.0.6.jar
