+++ adaptive-media-image-rest/bnd.bnd
@@ -1,0 +1,16 @@
+Bundle-Name: Liferay Adaptive Media Image REST
+Bundle-SymbolicName: com.liferay.adaptive.media.image.rest
+Bundle-Version: 1.0.0
+ConfigurationPath: /configuration
+Import-Package:\
+	!com.fasterxml.jackson.*,\
+	\
+	!com.google.gson,\
+	\
+	!javassist.*,\
+	\
+	!org.apache.commons.vfs2.*,\
+	*
+Include-Resource: configuration=src/main/resources/configuration
+Liferay-Releng-Module-Group-Description:
+Liferay-Releng-Module-Group-Title: Adaptive Media
