+++ adaptive-media-image-rest/build.gradle
