+++ adaptive-media-image-rest/src/main/java/com/liferay/adaptive/media/image/rest/internal/provider/CompanyContextProvider.java
@@ -1,0 +1,64 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.adaptive.media.image.rest.internal.provider;
+
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.model.Company;
+import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.util.PortalUtil;
+
+import javax.servlet.http.HttpServletRequest;
+
+import javax.ws.rs.ext.Provider;
+
+import org.apache.cxf.jaxrs.ext.ContextProvider;
+import org.apache.cxf.message.Message;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Carlos Sierra Andrés
+ */
+@Component(immediate = true, service = CompanyContextProvider.class)
+@Provider
+public class CompanyContextProvider implements ContextProvider<Company> {
+
+	@Override
+	public Company createContext(Message message) {
+		HttpServletRequest request =
+			(HttpServletRequest)message.getContextualProperty("HTTP.REQUEST");
+
+		try {
+			return PortalUtil.getCompany(request);
+		}
+		catch (PortalException pe) {
+			if (_log.isWarnEnabled()) {
+				_log.warn(pe);
+			}
+
+			return null;
+		}
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		CompanyContextProvider.class);
+
+	@Reference
+	private CompanyLocalService _companyLocalService;
+
+}
