+++ adaptive-media-image-rest/src/main/resources/configuration/com.liferay.portal.remote.rest.extender.configuration.RestExtenderConfiguration-adaptivemediajaxrs
@@ -1,0 +1,2 @@
+contextPaths=/adaptive-media-rest
+jaxRsApplicationFilterStrings=(component.name=com.liferay.adaptive.media.image.rest.internal.application.ImageAdaptiveMediaJaxRsApplication)
